cloud_functions/coingecko_all_coins/requirements.txt
functions-framework==3.*
requests==2.31.0
google-cloud-bigquery==3.10.1
google-cloud-storage==2.11.0
dreams_core==0.2.22